+++ modules/social-activity-customizer/social-activity-customizer-api/src/main/java/com/liferay/social/activity/customizer/service/persistence/CustomSocialActivitySetFinder.java
@@ -25,13 +25,13 @@ public interface CustomSocialActivitySetFinder {
 
 	public int countByU_C(long userId, long classNameId);
 
-	public int countByU_C_T(long userId, long classNameId, long[] types);
+	public int countByU_C_T(long userId, long classNameId, int[] types);
 
 	public java.util.List<com.liferay.social.kernel.model.SocialActivitySet>
 		findByU_C(long userId, long classNameId, int begin, int end);
 
 	public java.util.List<com.liferay.social.kernel.model.SocialActivitySet>
 		findByU_C_T(
-			long userId, long classNameId, long[] types, int begin, int end);
+			long userId, long classNameId, int[] types, int begin, int end);
 
 }
